feat: 토큰 api 변경 /api/managers/ -> /api/tokens, 복호화 api 구현
